requirements.txt
flask==2.0.1
python-dotenv==0.19.0
flask-cors==3.0.10
google-generativeai==0.3.0
requests==2.31.0
